Change (BW)Rated elements to use TokenBuckets instead of GapRate
The biggest functional change here (by far) is that these elements now
have a 1 second burst.  This means that a BWRatedUnqueue element
configured at 1MBps will unqueue 1MB as quickly as it possibly can
(because the TokenBucket will start out with 1MB worth of tokens), and
after that, it will average 1MBps.  However, the elements should
deliver much more reliable throughput, especially in cases where there
is some bursty CPU behavior.

The other change is that these elements all now include timers which
they are willing to use, so they should not ever end up pegging the
CPU.

This also greatly improves performance.  I believe that this is
because of TokenBucket's relative efficiency to GapRate.
  cat myconfig
    InfiniteSource(LIMIT 4000000, TIMESTAMP false, STOP true)
    -> Queue(CAPACITY 100)
    -> RatedUnqueue(4000000000)
    -> Discard
  time ./userlevel/click myconfig
    real		      0m6.319s
    user		      0m3.212s
    sys		      0m3.088s
  time ./userlevel/click myconfig
    real		      0m1.729s
    user		      0m1.200s
    sys		      0m0.520s

Signed-off-by: Cliff Frey <cliff@meraki.com>