Each thread gets its own TimerSet.
Each timer, therefore, is associated with exactly one RouterThread.

This is useful for a couple reasons.  First, we can prevent a Timer from
going off in parallel with a Task, by placing the Timer and the Task on the
same thread (with StaticThreadSched, for example).  Second, it reduces
TimerSet lock contention: the NullTask benchmark can run 4x faster (on
read.cs.ucla.edu, and only sometimes, but still).

Signed-off-by: Eddie Kohler <ekohler@gmail.com>