click/llrpc: add exclusive flag
We have some LLRPCs that are used instead of normal handlers so that no Args
parsing has to be done.  However, these LLRPCs can modify shared data
structures or push into the router.  This is a problem since LLRPCs
currently do not block or unblock the router.

This change adds a new "exclusive" flag that specifically does stop the
router, which should be mutually exclusive with "safe".  The letter E was
chosen over X because _CLICK_IOX is already a macro.